latest version

+++ bookstore/models.py
@@ -3,6 +3,7 @@ from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.models import User
 from django.contrib.auth import get_user_model
 from django.urls import reverse
+from django.template.defaultfilters import filesizeformat
 
 
 # User = get_user_model()
@@ -17,33 +18,38 @@ class User(AbstractUser):
         swappable = 'AUTH_USER_MODEL'
 
 
+
+class Category(models.Model):
+    name = models.CharField(max_length=100, unique=True)
+    description = models.TextField(blank=True)
+
+    def __str__(self):
+        return self.name
+
+
+# Modify the Book model to include a category
 class Book(models.Model):
     title = models.CharField(max_length=100)
     author = models.CharField(max_length=100)
-    year = models.CharField(max_length=100)
-    publisher = models.CharField(max_length=200)
-    desc = models.CharField(max_length=1000)
     uploaded_by = models.CharField(max_length=100, null=True, blank=True)
     user_id = models.CharField(max_length=100, null=True, blank=True)
     pdf = models.FileField(upload_to='bookapp/zips/')
-    cover = models.ImageField(upload_to='bookapp/covers/', null=True, blank=True)
+    # Add a foreign key to Category
+    category = models.ForeignKey(Category, on_delete=models.SET_NULL, null=True)
 
     def __str__(self):
         return self.title
 
     def delete(self, *args, **kwargs):
         self.pdf.delete()
-        self.cover.delete()
+        if hasattr(self, 'cover'):
+            self.cover.delete()
         super().delete(*args, **kwargs)
+    @property
+    def formatted_file_size(self):
+        return filesizeformat(self.pdf.size)
 
 
-class Chat(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    message = models.TextField()
-    posted_at = models.DateTimeField(auto_now=True, null=True)
-
-    def __str__(self):
-        return str(self.message)
 
 
 class DeleteRequest(models.Model):
